name changes to all poly files and tests
