+++ .gitignore
@@ -87,16 +87,16 @@ symengine/tests/basic/test_series_expansion_UP
 symengine/tests/basic/test_series_expansion_URatP
 symengine/tests/basic/test_series_expansion_URatF
 symengine/tests/basic/test_polynomial_multivariate
-symengine/tests/polynomial/test_mexprpoly
-symengine/tests/polynomial/test_mintpoly
-symengine/tests/polynomial/test_uexprpoly
-symengine/tests/polynomial/test_uintpoly
-symengine/tests/polynomial/test_uratpoly
-symengine/tests/polynomial/test_uintpoly_flint
-symengine/tests/polynomial/test_uratpoly_flint
-symengine/tests/polynomial/test_uintpoly_piranha
-symengine/tests/polynomial/test_uratpoly_piranha
-symengine/tests/polynomial/test_basic_conversions
+symengine/tests/polys/test_mpoly_expr_symengine
+symengine/tests/polys/test_mpoly_int_symengine
+symengine/tests/polys/test_upoly_expr_symengine
+symengine/tests/polys/test_upoly_int_symengine
+symengine/tests/polys/test_upoly_rat_symengine
+symengine/tests/polys/test_upoly_int_flint
+symengine/tests/polys/test_upoly_rat_flint
+symengine/tests/polys/test_upoly_int_piranha
+symengine/tests/polys/test_upoly_rat_piranha
+symengine/tests/polys/test_basic_conversions
 symengine/tests/basic/test_fields
 symengine/tests/basic/test_infinity
 symengine/tests/logic/test_logic
