+++ symengine/fields.cpp
@@ -71,7 +71,7 @@ GaloisField::from_vec(const RCP<const Basic> &var,
                                        GaloisFieldDict::from_vec(v, modulo));
 }
 
-RCP<const GaloisField> GaloisField::from_uintpoly(const UIntPoly &a,
+RCP<const GaloisField> GaloisField::from_upoly_int_symengine(const UIntPoly &a,
                                                   const integer_class &modulo)
 {
     GaloisFieldDict wrapper(a.get_poly().get_dict(), modulo);
