+++ symengine/polys/upoly_base.h
