+++ symengine/tests/CMakeLists.txt
@@ -7,7 +7,7 @@ add_subdirectory(basic)
 add_subdirectory(ntheory)
 add_subdirectory(matrix)
 add_subdirectory(printing)
-add_subdirectory(polynomial)
+add_subdirectory(polys)
 add_subdirectory(eval)
 add_subdirectory(cwrapper)
 add_subdirectory(expression)
