+++ symengine/tests/basic/test_fields.cpp
@@ -46,11 +46,11 @@ TEST_CASE("Constructor of GaloisField : Basic", "[basic]")
 
     RCP<const UIntPoly> UP
         = UIntPoly::from_dict(x, {{0, 3_z}, {1, 4_z}, {2, 5_z}});
-    U = GaloisField::from_uintpoly(*UP, 5_z);
+    U = GaloisField::from_upoly_int_symengine(*UP, 5_z);
     REQUIRE(U->__str__() == "4*x + 3");
 
     UP = UIntPoly::from_dict(x, {{0, 10_z}, {1, 7_z}, {2, 9_z}});
-    U = GaloisField::from_uintpoly(*UP, 7_z);
+    U = GaloisField::from_upoly_int_symengine(*UP, 7_z);
     REQUIRE(U->__str__() == "2*x**2 + 3");
 }
