+++ symengine/tests/polys/test_upoly_int_flint.cpp
@@ -1,9 +1,9 @@
 #include "catch.hpp"
 #include <chrono>
 
-#include <symengine/polys/uintpoly.h>
-#include <symengine/polys/uintpoly_flint.h>
-#include <symengine/polys/uintpoly_piranha.h>
+#include <symengine/polys/upoly_int_symengine.h>
+#include <symengine/polys/upoly_flint.h>
+#include <symengine/polys/upoly_piranha.h>
 #include <symengine/pow.h>
 #include <symengine/dict.h>
 #include <symengine/add.h>
