+++ symengine/tests/polys/test_upoly_int_symengine.cpp
@@ -3,11 +3,11 @@
 
 #include <symengine/mul.h>
 #include <symengine/add.h>
-#include <symengine/polys/uintpoly.h>
+#include <symengine/polys/upoly_int_symengine.h>
 #include <symengine/pow.h>
 #include <symengine/symengine_exception.h>
-#include <symengine/polys/uintpoly_piranha.h>
-#include <symengine/polys/uintpoly_flint.h>
+#include <symengine/polys/upoly_piranha.h>
+#include <symengine/polys/upoly_flint.h>
 
 #ifdef HAVE_SYMENGINE_PIRANHA
 using SymEngine::UIntPolyPiranha;
