+++ symengine/tests/polys/test_upoly_rat_piranha.cpp
@@ -1,12 +1,12 @@
 #include "catch.hpp"
 #include <chrono>
 
-#include <symengine/polys/uintpoly_piranha.h>
+#include <symengine/polys/upoly_piranha.h>
 #include <symengine/pow.h>
 #include <symengine/symengine_exception.h>
-#include <symengine/polys/uintpoly_flint.h>
-#include <symengine/polys/uratpoly.h>
-#include <symengine/polys/uintpoly.h>
+#include <symengine/polys/upoly_flint.h>
+#include <symengine/polys/upoly_rat_symengine.h>
+#include <symengine/polys/upoly_int_symengine.h>
 
 #ifdef HAVE_SYMENGINE_FLINT
 using SymEngine::URatPolyFlint;
